WIP removing old workflows
