+++ backend/models/__init__.py
@@ -1,6 +1,7 @@
 from .auth import *
 from .base import *
 from .cans import *
+from .change_requests import *
 from .document import *
 from .events import *
 from .history import *
@@ -10,4 +11,3 @@ from .procurement_shops import *
 from .projects import *
 from .users import *
 from .vendors import *
-from .workflows import *
