--- backend/ops_api/ops/resources/package.py
@@ -1,44 +1,0 @@
-from flask import Response
-
-from models.base import BaseModel
-from ops_api.ops.auth.auth_types import Permission, PermissionType
-from ops_api.ops.auth.decorators import is_authorized
-from ops_api.ops.base_views import BaseItemAPI, BaseListAPI
-
-ENDPOINT_STRING = "/package"
-
-
-class PackageItemAPI(BaseItemAPI):
-    def __init__(self, model: BaseModel):
-        super().__init__(model)
-
-    @is_authorized(PermissionType.GET, Permission.WORKFLOW)
-    def get(self, id: int) -> Response:
-        return self._get_item_with_try(id)
-
-
-class PackageListAPI(BaseListAPI):
-    def __init__(self, model: BaseModel):
-        super().__init__(model)
-
-    @is_authorized(PermissionType.GET, Permission.WORKFLOW)
-    def get(self) -> Response:
-        return super().get()
-
-
-class PackageSnapshotItemAPI(BaseItemAPI):
-    def __init__(self, model: BaseModel):
-        super().__init__(model)
-
-    @is_authorized(PermissionType.GET, Permission.WORKFLOW)
-    def get(self, id: int) -> Response:
-        return self._get_item_with_try(id)
-
-
-class PackageSnapshotListAPI(BaseListAPI):
-    def __init__(self, model: BaseModel):
-        super().__init__(model)
-
-    @is_authorized(PermissionType.GET, Permission.WORKFLOW)
-    def get(self) -> Response:
-        return super().get()
