--- backend/ops_api/ops/resources/workflow_template.py
@@ -1,26 +1,0 @@
-from flask import Response
-
-from models.base import BaseModel
-from ops_api.ops.auth.auth_types import Permission, PermissionType
-from ops_api.ops.auth.decorators import is_authorized
-from ops_api.ops.base_views import BaseItemAPI, BaseListAPI
-
-ENDPOINT_STRING = "/workflow-template"
-
-
-class WorkflowTemplateItemAPI(BaseItemAPI):
-    def __init__(self, model: BaseModel):
-        super().__init__(model)
-
-    @is_authorized(PermissionType.GET, Permission.WORKFLOW)
-    def get(self, id: int) -> Response:
-        return self._get_item_with_try(id)
-
-
-class WorkflowTemplateListAPI(BaseListAPI):
-    def __init__(self, model: BaseModel):
-        super().__init__(model)
-
-    @is_authorized(PermissionType.GET, Permission.WORKFLOW)
-    def get(self) -> Response:
-        return super().get()
