+++ backend/ops_api/ops/schemas/agreements.py
@@ -29,7 +29,7 @@ class AgreementData(Schema):
     project_id = fields.Integer(allow_none=True)
     procurement_shop_id = fields.Integer(allow_none=True)
     notes = fields.String(allow_none=True)
-    procurement_tracker_workflow_id = fields.Integer(allow_none=True)
+    procurement_tracker_id = fields.Integer(allow_none=True)
 
 
 class ContractAgreementData(AgreementData):
