+++ backend/ops_api/ops/urls.py
@@ -60,14 +60,6 @@ from ops_api.ops.views import (
     USERS_ITEM_API_VIEW_FUNC,
     USERS_LIST_API_VIEW_FUNC,
     VERSION_API_VIEW_FUNC,
-    WORKFLOW_INSTANCE_ITEM_API_VIEW_FUNC,
-    WORKFLOW_INSTANCE_LIST_API_VIEW_FUNC,
-    WORKFLOW_STEP_INSTANCE_ITEM_API_VIEW_FUNC,
-    WORKFLOW_STEP_INSTANCE_LIST_API_VIEW_FUNC,
-    WORKFLOW_STEP_TEMPLATE_ITEM_API_VIEW_FUNC,
-    WORKFLOW_STEP_TEMPLATE_LIST_API_VIEW_FUNC,
-    WORKFLOW_TEMPLATE_ITEM_API_VIEW_FUNC,
-    WORKFLOW_TEMPLATE_LIST_API_VIEW_FUNC,
 )
 
 # Ideas from Flask docs: https://flask.palletsprojects.com/en/2.2.x/views/#method-dispatching-and-apis
@@ -276,38 +268,6 @@ def register_api(api_bp: Blueprint) -> None:
         view_func=NOTIFICATIONS_ITEM_API_VIEW_FUNC,
     )
 
-    api_bp.add_url_rule(
-        "/workflow-instance/",
-        view_func=WORKFLOW_INSTANCE_LIST_API_VIEW_FUNC,
-    )
-    api_bp.add_url_rule(
-        "/workflow-instance/<int:id>",
-        view_func=WORKFLOW_INSTANCE_ITEM_API_VIEW_FUNC,
-    )
-    api_bp.add_url_rule(
-        "/workflow-step-instance/",
-        view_func=WORKFLOW_STEP_INSTANCE_LIST_API_VIEW_FUNC,
-    )
-    api_bp.add_url_rule(
-        "/workflow-step-instance/<int:id>",
-        view_func=WORKFLOW_STEP_INSTANCE_ITEM_API_VIEW_FUNC,
-    )
-    api_bp.add_url_rule(
-        "/workflow-template/",
-        view_func=WORKFLOW_TEMPLATE_LIST_API_VIEW_FUNC,
-    )
-    api_bp.add_url_rule(
-        "/workflow-template/<int:id>",
-        view_func=WORKFLOW_TEMPLATE_ITEM_API_VIEW_FUNC,
-    )
-    api_bp.add_url_rule(
-        "/workflow-step-template/",
-        view_func=WORKFLOW_STEP_TEMPLATE_LIST_API_VIEW_FUNC,
-    )
-    api_bp.add_url_rule(
-        "/workflow-step-template/<int:id>",
-        view_func=WORKFLOW_STEP_TEMPLATE_ITEM_API_VIEW_FUNC,
-    )
     api_bp.add_url_rule(
         "/services-components/<int:id>",
         view_func=SERVICES_COMPONENT_ITEM_API_VIEW_FUNC,
