+++ backend/ops_api/tests/ops/workflows/test_change_requests.py
@@ -19,51 +19,11 @@ from models import (
     ContractType,
     Division,
     ServiceRequirementType,
-    WorkflowStepTemplate,
-    WorkflowStepType,
-    WorkflowTemplate,
 )
 from ops_api.ops.resources.agreement_history import find_agreement_histories
-from ops_api.ops.utils.procurement_workflow_helper import delete_procurement_workflow
+from ops_api.ops.utils.procurement_tracker_helper import delete_procurement_tracker
 
 test_no_perms_user_id = 506
-
-
-@pytest.mark.usefixtures("app_ctx")
-def test_workflow_template_retrieve(auth_client, loaded_db):
-    workflow_template = loaded_db.get(WorkflowTemplate, 1)
-
-    assert workflow_template is not None
-    assert workflow_template.name == "Basic Approval"
-    assert workflow_template.steps is not None
-
-
-@pytest.mark.usefixtures("app_ctx")
-def test_workflow_step_template_retrieve(auth_client, loaded_db):
-    workflow_step_template = loaded_db.get(WorkflowStepTemplate, 1)
-
-    assert workflow_step_template is not None
-    assert workflow_step_template.name == "Initial Review"
-    assert workflow_step_template.workflow_type == WorkflowStepType.APPROVAL
-    assert workflow_step_template.index == 0
-    assert workflow_step_template.step_approvers is not None
-
-
-@pytest.mark.usefixtures("app_ctx")
-@pytest.mark.usefixtures("loaded_db")
-def test_get_workflow_template_by_id(auth_client):
-    response = auth_client.get("/api/v1/workflow-template/1")
-    assert response.status_code == 200
-    assert response.json["id"] == 1
-
-
-@pytest.mark.usefixtures("app_ctx")
-@pytest.mark.usefixtures("loaded_db")
-def test_get_workflow_step_template_by_id(auth_client):
-    response = auth_client.get("/api/v1/workflow-step-template/1")
-    assert response.status_code == 200
-    assert response.json["id"] == 1
-
 
 # ---=== CHANGE REQUESTS ===---
 
@@ -578,7 +538,7 @@ def test_status_change_request_creates_procurement_workflow(
     assert bli.agreement.procurement_tracker_workflow_id is not None
 
     # cleanup
-    delete_procurement_workflow(bli.agreement_id)
+    delete_procurement_tracker(bli.agreement_id)
     loaded_db.delete(bli)
     loaded_db.delete(agreement)
     loaded_db.commit()
